Move redis as stand-alone module
