+++ services/web/server/src/simcore_service_webserver/projects/project_lock.py
@@ -5,10 +5,7 @@ from aiohttp import web
 from models_library.projects import ProjectID
 from models_library.projects_state import Owner, ProjectLocked, ProjectStatus
 
-from ..resource_manager.redis import (
-    get_redis_lock_manager,
-    get_redis_lock_manager_client,
-)
+from ..redis import get_redis_lock_manager, get_redis_lock_manager_client
 from ..users_api import UserNameDict
 
 PROJECT_REDIS_LOCK_KEY: str = "project:{}"
