+++ services/web/server/src/simcore_service_webserver/resource_manager/module_setup.py
@@ -9,11 +9,15 @@
 import logging
 
 from aiohttp import web
-from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
+from servicelib.aiohttp.application_setup import (
+    AlreadyInitializedError,
+    ModuleCategory,
+    app_module_setup,
+)
 
+from ..redis import setup_redis
 from .config import APP_CLIENT_SOCKET_REGISTRY_KEY, APP_RESOURCE_MANAGER_TASKS_KEY
 from .garbage_collector import setup_garbage_collector
-from .redis import setup_redis_client
 from .registry import RedisResourceRegistry
 
 logger = logging.getLogger(__name__)
@@ -32,7 +36,11 @@ def setup_resource_manager(app: web.Application) -> bool:
     # ---------------------------------------------
 
     app[APP_RESOURCE_MANAGER_TASKS_KEY] = []
-    setup_redis_client(app)
+    try:
+        setup_redis(app)
+    except AlreadyInitializedError as err:
+        logger.info("Skips setting up redis client: %s", err)
+
     app[APP_CLIENT_SOCKET_REGISTRY_KEY] = RedisResourceRegistry(app)
     setup_garbage_collector(app)
     return True
