+++ services/web/server/src/simcore_service_webserver/resource_manager/registry.py
@@ -19,8 +19,8 @@ from typing import Dict, List, Tuple
 import aioredis
 from aiohttp import web
 
+from ..redis import get_redis_client
 from .config import APP_CLIENT_SOCKET_REGISTRY_KEY
-from .redis import get_redis_client
 
 log = logging.getLogger(__name__)
