+++ services/web/server/src/simcore_service_webserver/studies_access.py
@@ -23,8 +23,8 @@ from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setu
 
 from ._constants import INDEX_RESOURCE_NAME
 from .login.decorators import login_required
+from .redis import get_redis_lock_manager
 from .resource_manager.config import GUEST_USER_RC_LOCK_FORMAT
-from .resource_manager.redis import get_redis_lock_manager
 from .security_api import is_anonymous, remember
 from .storage_api import copy_data_folders_from_project
 from .utils import compose_error_msg
